libtest: Split formatters into modules
The usage of `::*` is a bit ugly, but easy to fix when the code is less
often changing.
